// src/frontend/packages/core/src/cf-favourite-types.ts
import { IFavoriteMetadata } from '../../store/src/types/user-favorites.types';

export interface ISpaceFavMetadata extends IFavoriteMetadata {
  guid: string;
  orgGuid: string;
  name: string;
  cfGuid: string;
}

export interface IOrgFavMetadata extends IFavoriteMetadata {
  guid: string;
  status: string;
  name: string;
  cfGuid: string;
}


export interface IAppFavMetadata extends IFavoriteMetadata {
  guid: string;
  cfGuid: string;
  name: string;
}